FB-16 #comment Development (Split Service and repository. Rename factory to mapper)
