+++ src/main/java/no/fint/xledger/graphql/ObjectKinds.java
@@ -1,0 +1,6 @@
+package no.fint.xledger.graphql;
+
+public class ObjectKinds {
+    public static final int MERVERDIAVGIFT = 14;
+    public static final int SALGSORDREGRUPPE = 156;
+}
