+++ src/main/java/no/fint/xledger/okonomi/fakturagrunnlag/UpdateFakturagrunnlagHandler.java
@@ -1,4 +1,4 @@
-package no.fint.xledger.okonomi;
+package no.fint.xledger.okonomi.fakturagrunnlag;
 
 import no.fint.xledger.handler.Handler;
 import no.fint.event.model.Event;
